requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
appnope==0.1.4
asttokens==3.0.0
attrs==24.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
comm==0.2.2
debugpy==1.8.11
decorator==5.1.1
executing==2.1.0
frozenlist==1.5.0
idna==3.10
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
lightstreamer-client-lib==2.2.0
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
numpy==2.2.0
opencv-python==4.10.0.84
packaging==24.2
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pip-system-certs==4.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyzmq==26.2.0
requests==2.32.3
six==1.17.0
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
websockets==14.1
wrapt==1.17.0
yarl==1.18.3
